refactor: import 경로 구체화 alias 설정 및 파일명 변경
