+++ src/shared/components/index.ts
@@ -2,10 +2,12 @@ export * from './lucide-icon'
 // export * from './timepicker'
 export * from './buttons'
 export * from './calender'
+export * from './chatbot'
 export * from './check-icon'
 export * from './datepicker'
 export * from './dropdown'
 export * from './errortext'
+export * from './feedback'
 export * from './header'
 export * from './input'
 export * from './leaf-reward'
@@ -16,4 +18,3 @@ export * from './select'
 export * from './slidearea'
 export * from './switchtap'
 export * from './toast'
-export * from './feedback'
