+++ src/shared/components/timepicker/dropdown/index.ts
@@ -1,1 +1,1 @@
-export * from './time-dropdown'
+export * from './dropdown'
